operators/command_source: Fix empty directory removal

+++ operators/command_source.zsh
@@ -14,11 +14,14 @@ done
 
 source "${units}/sn_auth.zsh"
 
-user_data=($(ls -1 "${user_config}"))
-
-if [[ ${#user_data} -eq 0 ]]
+if [[ -d "${user_config}" ]]
 then
-    rmdir "${user_config}"
+    user_data=($(ls -1 "${user_config}"))
+
+    if [[ ${#user_data} -eq 0 ]]
+    then
+        rmdir "${user_config}"
+    fi
 fi
 
 rmdir "${user_config}.lock"
